MOD: Rename packages: controller to controllers and model to models
