+++ src/courseView/CourseSearchMenuView.java
@@ -2,11 +2,11 @@ package courseView;
 
 import java.util.ArrayList;
 
-import controller.CourseController;
-import controller.MainController;
-import model.Course;
-import model.Teacher;
-import model.User;
+import controllers.CourseController;
+import controllers.MainController;
+import models.Course;
+import models.Teacher;
+import models.User;
 import tools.Alerts;
 import tools.Cnsl;
